Load the javascript as a static file.
