+++ django_js_error_hook/urls.py
@@ -1,8 +1,6 @@
 from django.urls import re_path
-from .views import js_error_view, utils_js
-
+from .views import js_error_view
 
 urlpatterns = [
     re_path("^$", js_error_view, name="js-error-handler"),
-    re_path("^utils.js$", utils_js, name="js-error-handler-js"),
 ]
